requirements.txt
# Core dependencies for Privacy Policy Analyzer
openai>=1.0.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional dependencies for development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0